dts: pinctrl: Introduce pincfg-node
Many data items that are represented in a pin configuration node are
common and generic. Pin control bindings should use the properties
defined below where they are applicable; not all of these properties
are relevant or useful for all hardware or binding structures. Each
individual binding document should state which of these generic
properties, if any, are used, and the structure of the DT nodes that
contain these properties.

This is based on Linux, documentation:
https://www.kernel.org/doc/Documentation/devicetree/bindings/pinctrl/pincfg-node.yaml

Add pincfg-node.yaml properties and refactor atmel,sam-pinctrl.yaml
to use these properties.

Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>