boards: arm: cy8ckit_062_ble: m0: Add serial pinctrl and int
Add pinctrl and interrupt information for uart-5.  The uart-5 is the
default serial communication on the kit.  It uses SCB[5] and the NVIC
line 21 as interrupt source on Cortex-M0+ cpu.  The line 21 is an
arbitrary value and users can select any value.  The SCB[5] shares
pins with Arduino headers D0/D1 and it is wired with the on-board USB
Serial converter.  This means, if use wants use an external shield it
must rework the board.

Signed-off-by: Gerson Fernando Budke <gerson.budke@atl-electronics.com>