boards: arm: cy8ckit_062_wifi_bt: m0: Add serial pinctrl and int
Add pinctrl and interrupt information for uart-6.  The uart-6 is the
default serial communication on the kit.  It uses SCB[6] and the NVIC
line 21 as interrupt source on Cortex-M0+ cpu.  The line 21 is an
arbitrary value and users can select any value.  The SCB[6] shares
pins with Arduino headers D8/D9.  It my necessary an USB serial
converter to sample signals and enable Zephyr console on PC.

Signed-off-by: Gerson Fernando Budke <gerson.budke@atl-electronics.com>