trusted-firmware-m: Define header and trailer options
The current behaviour when signing an image is to set always --pad and
--pad-header for all images unless TFM_USE_NS_APP is set. This does not
allow easy creation of a signed image for FOTA applications. Rewrite
the PAD parameter into HEADER and TRAILER to easy setup more signing
options.

The other important reason to do this change is that NS image when
signed without --pad run in the hardware. However, it does not perform
the MCUboot test and the FWU never upgrades the image. This will fix
the NS image signing process to correct allow TF-M FWU using the PSA
API functions.

Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>