trusted-firmware-m: Set --confirm when signing
The current behaviour from signing an image add --pad but do not confirm
the image. This seems to be a mistake because the user should inspect
the image status in the Firmware Upgrade software. This means that if an
image is not --confirmed the FSM can not infer correct states. This set
the image as confirmed to fix this issue.

Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>