trusted-firmware-m: Create multi image bin files
One fundamental use of trusted-firmware-m is to allow IoT applications
to have security and the firmware upgrade FOTA is almost 100% mandatory
in these applications. The current state of files signing process do
not produce the necessary binaries to use with multi image S/NS FWU
since the hex images are not suitable for this use case. This
introduces the missing signed binaries files to be used by FWU
partition. The changes were tested in multi image FWU scenarios and
single image can be easily extended in the future.

Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>