trusted-firmware-m: Create bin files to FOTA
One fundamental use of trusted-firmware-m is to allow IoT applications
to have security and the firmware upgrade is almost 100% mandatory in
these applications. The current state of files signing process do not
produce the necessary binaries to use with FWU because the hex images
are not suitable for this use case. This introduces the missing signed
binaries files to be used by FWU partition.

Signed-off-by: BUDKE Gerson Fernando <gerson.budke@leica-geosystems.com>